backend/requirements.txt
fastapi[standard]
matplotlib
pandas
seaborn
scikit-learn
openpyxl
requests
numpy==1.26.4
biopython
packaging
psutil